dockerpinger/services/ponger/requirements.txt
kafka-python==1.4.4
